Updated distribution package in greedy action head and moved types out of iql file.
